Extract fetchCoach to a single place
