+++ client/coach_view/coach_view.html
@@ -27,8 +27,11 @@
                     <header><h2><a id="create_swimmer_link" href="../swimmer_create/swimmer_create.html"></a></h2></header>
                     <section>
                         <div class="details-grid">
-                            <label for="name">Name:</label><span id="name">name</span>
-                            <label for="level">Level:</label><span id="level">level</span>
+                            <label for="name">Name:</label>
+                            <input class="ui-control ui-input" id="name" name="name" readonly/>
+
+                            <label for="level">Level:</label>
+                            <input class="ui-control ui-input" id="level" name="level" readonly/>
                         </div>
                     </section>
                     <section>
